Further organize Line; Fix delimiter test
