+++ DiffFormatter/App/Info.plist
@@ -9,6 +9,6 @@
 	<key>CFBundleShortVersionString</key>
 	<string>0.0.10</string>
 	<key>CFBundleVersion</key>
-	<string>85</string>
+	<string>88</string>
 </dict>
 </plist>
