+++ DiffFormatterTests/Config/custom_delimiter_config.json
@@ -45,5 +45,6 @@
             "left": "**❲**",
             "right": "**❳**"
         }
-    }
+    },
+    "repo_base_url": "https://github.com/citadel-of-ricks/C-137"
 }
